Allow inventory cost estimates to run on just a subset of dates.
Change the inventory summarization code so that it can rewrite the data
for just more recent dates (this will be important if we want to keep
data from old financial periods stable).  However, still do a
computation over a slightly longer time period, so that the estimates
will tend to converge to the same value instead of having a
discontinuity at the start of the period.

+++ django/chezbob/util/inventory.py
@@ -159,16 +159,28 @@ def generate_inventory_report(start, end=None):
         date += ONE_DAY
 
 @django.db.transaction.commit_manually
-def summary(start, end=None, fp=None):
-    print "Generating Inventory Summary..."
-
+def summary(start, end=None, commit_start=None):
     date = None
     losses = 0.0
     inventory_value = 0.0
 
     def commit_day():
-        print "%s: inventory=%.2f, shrinkage=%.2f" \
-            % (date, inventory_value, losses)
+        cursor.execute("""SELECT value FROM finance_inventory_summary
+                          WHERE date='%s'""",
+                       (date,))
+        old_value = 0.0
+        for r in cursor.fetchall():
+            old_value = r[0]
+
+        if date == commit_start:
+            print "--- COMMITTING ---"
+
+        print "%s: inventory=%.2f shrinkage=%.2f [old_value=%.2f delta=%.2f]" \
+            % (date, inventory_value, losses,
+               old_value, inventory_value - old_value)
+
+        if commit_day is not None and date < commit_start:
+            return
 
         cursor.execute("DELETE FROM finance_inventory_summary WHERE date='%s'",
                        (date,))
@@ -177,7 +189,13 @@ def summary(start, end=None, fp=None):
                           VALUES ('%s', %s, %s)""",
                        (date, inventory_value, losses))
 
+    print "Gathering price data..."
+    first_result_seen = False
     for inv in generate_inventory_report(start, end):
+        if not first_result_seen:
+            print "Summarizing inventory data..."
+            first_result_seen = True
+
         if inv['date'] != date:
             if date is not None: commit_day()
             date = inv['date']
